requirements.txt
gitpython>=3.1.40
pandas>=2.1.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0
